requirements.txt
Django==3.2
django-crispy-forms
crispy-bootstrap5
python-decouple
djangorestframework
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
django-rest-passwordreset
django-mail-templated
black
flake8
pytest
pytest-django